requirements.txt
customtkinter>=5.2
yt-dlp>=2024.3
pillow>=10.0
requests>=2.31
telethon>=1.34
cryptg>=0.4
hachoir>=3.2
tkinterdnd2>=0.3
lxml>=4.9